+++ Forms/MuxerForm.vb
@@ -217,6 +217,7 @@ Public Class MuxerForm
         '
         'dgvSubtitles
         '
+        Me.dgvSubtitles.AllowDrop = True
         Me.dgvSubtitles.Anchor = CType((((System.Windows.Forms.AnchorStyles.Top Or System.Windows.Forms.AnchorStyles.Bottom) _
             Or System.Windows.Forms.AnchorStyles.Left) _
             Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)