+++ Forms/eac3toForm.vb
@@ -3,6 +3,7 @@ Imports System.ComponentModel
 Imports System.Globalization
 Imports System.Text.RegularExpressions
 Imports System.Threading.Tasks
+
 Imports StaxRip.UI
 
 Public Class eac3toForm
@@ -20,7 +21,7 @@ Public Class eac3toForm
     End Sub
     Friend WithEvents CommandLink1 As StaxRip.UI.CommandLink
     Friend WithEvents cbVideoOutput As System.Windows.Forms.ComboBox
-    Friend WithEvents Label5 As System.Windows.Forms.Label
+    Friend WithEvents laVideo As System.Windows.Forms.Label
     Friend WithEvents cmdlOptions As StaxRip.CommandLineControl
     Friend WithEvents bnBrowse As StaxRip.UI.ButtonEx
     Friend WithEvents bnCancel As StaxRip.UI.ButtonEx
@@ -28,17 +29,17 @@ Public Class eac3toForm
     Friend WithEvents lvAudio As ListViewEx
     Friend WithEvents lvSubtitles As StaxRip.UI.ListViewEx
     Friend WithEvents cbVideoStream As System.Windows.Forms.ComboBox
-    Friend WithEvents Label8 As System.Windows.Forms.Label
+    Friend WithEvents laStream As System.Windows.Forms.Label
     Friend WithEvents flpAudioLinks As System.Windows.Forms.FlowLayoutPanel
     Friend WithEvents flpSubtitleLinks As System.Windows.Forms.FlowLayoutPanel
     Friend WithEvents cbChapters As System.Windows.Forms.CheckBox
-    Friend WithEvents Label2 As System.Windows.Forms.Label
+    Friend WithEvents laTargetDir As System.Windows.Forms.Label
     Friend WithEvents tlpTarget As System.Windows.Forms.TableLayoutPanel
     Friend WithEvents gbAudio As System.Windows.Forms.GroupBox
     Friend WithEvents gbSubtitles As System.Windows.Forms.GroupBox
     Friend WithEvents cbAudioOutput As System.Windows.Forms.ComboBox
-    Friend WithEvents Label3 As System.Windows.Forms.Label
-    Friend WithEvents Label1 As System.Windows.Forms.Label
+    Friend WithEvents laOptions As System.Windows.Forms.Label
+    Friend WithEvents laOutput As System.Windows.Forms.Label
     Friend WithEvents bnMenu As StaxRip.UI.ButtonEx
     Friend WithEvents cms As ContextMenuStripEx
     Friend WithEvents bnAudioAll As ButtonEx
@@ -62,7 +63,7 @@ Public Class eac3toForm
         Me.components = New System.ComponentModel.Container()
         Me.cmdlOptions = New StaxRip.CommandLineControl()
         Me.cbVideoOutput = New System.Windows.Forms.ComboBox()
-        Me.Label5 = New System.Windows.Forms.Label()
+        Me.laVideo = New System.Windows.Forms.Label()
         Me.bnBrowse = New StaxRip.UI.ButtonEx()
         Me.bnCancel = New StaxRip.UI.ButtonEx()
         Me.bnOK = New StaxRip.UI.ButtonEx()
@@ -79,9 +80,9 @@ Public Class eac3toForm
         Me.bnAudioEnglish = New StaxRip.UI.ButtonEx()
         Me.bnAudioNative = New StaxRip.UI.ButtonEx()
         Me.cbVideoStream = New System.Windows.Forms.ComboBox()
-        Me.Label8 = New System.Windows.Forms.Label()
+        Me.laStream = New System.Windows.Forms.Label()
         Me.tlpTarget = New System.Windows.Forms.TableLayoutPanel()
-        Me.Label2 = New System.Windows.Forms.Label()
+        Me.laTargetDir = New System.Windows.Forms.Label()
         Me.teTempDir = New StaxRip.UI.TextEdit()
         Me.tlpBottom = New System.Windows.Forms.TableLayoutPanel()
         Me.bnMenu = New StaxRip.UI.ButtonEx()
@@ -90,8 +91,8 @@ Public Class eac3toForm
         Me.gbAudio = New System.Windows.Forms.GroupBox()
         Me.tlpAudio = New System.Windows.Forms.TableLayoutPanel()
         Me.tlpAudioOptions = New System.Windows.Forms.TableLayoutPanel()
-        Me.Label1 = New System.Windows.Forms.Label()
-        Me.Label3 = New System.Windows.Forms.Label()
+        Me.laOutput = New System.Windows.Forms.Label()
+        Me.laOptions = New System.Windows.Forms.Label()
         Me.cbAudioOutput = New System.Windows.Forms.ComboBox()
         Me.gbSubtitles = New System.Windows.Forms.GroupBox()
         Me.tlpSubtitles = New System.Windows.Forms.TableLayoutPanel()
@@ -113,10 +114,10 @@ Public Class eac3toForm
         'cmdlOptions
         '
         Me.cmdlOptions.Anchor = CType((System.Windows.Forms.AnchorStyles.Left Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
-        Me.cmdlOptions.Location = New System.Drawing.Point(520, 8)
+        Me.cmdlOptions.Location = New System.Drawing.Point(560, 8)
         Me.cmdlOptions.Margin = New System.Windows.Forms.Padding(3, 8, 3, 8)
         Me.cmdlOptions.Name = "cmdlOptions"
-        Me.cmdlOptions.Size = New System.Drawing.Size(488, 70)
+        Me.cmdlOptions.Size = New System.Drawing.Size(702, 70)
         Me.cmdlOptions.TabIndex = 5
         '
         'cbVideoOutput
@@ -124,36 +125,36 @@ Public Class eac3toForm
         Me.cbVideoOutput.Anchor = System.Windows.Forms.AnchorStyles.None
         Me.cbVideoOutput.DropDownStyle = System.Windows.Forms.ComboBoxStyle.DropDownList
         Me.cbVideoOutput.FormattingEnabled = True
-        Me.cbVideoOutput.Location = New System.Drawing.Point(129, 3)
+        Me.cbVideoOutput.Location = New System.Drawing.Point(169, 3)
         Me.cbVideoOutput.Name = "cbVideoOutput"
         Me.cbVideoOutput.Size = New System.Drawing.Size(206, 56)
         Me.cbVideoOutput.TabIndex = 2
         '
-        'Label5
+        'laVideo
         '
-        Me.Label5.Anchor = System.Windows.Forms.AnchorStyles.None
-        Me.Label5.AutoSize = True
-        Me.Label5.Location = New System.Drawing.Point(3, 7)
-        Me.Label5.Name = "Label5"
-        Me.Label5.Size = New System.Drawing.Size(120, 48)
-        Me.Label5.TabIndex = 1
-        Me.Label5.Text = "Video:"
-        Me.Label5.TextAlign = System.Drawing.ContentAlignment.MiddleLeft
+        Me.laVideo.Anchor = System.Windows.Forms.AnchorStyles.None
+        Me.laVideo.AutoSize = True
+        Me.laVideo.Location = New System.Drawing.Point(3, 7)
+        Me.laVideo.Name = "laVideo"
+        Me.laVideo.Size = New System.Drawing.Size(160, 48)
+        Me.laVideo.TabIndex = 1
+        Me.laVideo.Text = " Output: "
+        Me.laVideo.TextAlign = System.Drawing.ContentAlignment.MiddleLeft
         '
         'bnBrowse
         '
         Me.bnBrowse.Anchor = System.Windows.Forms.AnchorStyles.None
-        Me.bnBrowse.Location = New System.Drawing.Point(959, 4)
+        Me.bnBrowse.Location = New System.Drawing.Point(1171, 4)
         Me.bnBrowse.Margin = New System.Windows.Forms.Padding(0)
-        Me.bnBrowse.Size = New System.Drawing.Size(70, 70)
+        Me.bnBrowse.Size = New System.Drawing.Size(100, 70)
         Me.bnBrowse.Text = "..."
         '
         'bnCancel
         '
         Me.bnCancel.Anchor = System.Windows.Forms.AnchorStyles.None
         Me.bnCancel.DialogResult = System.Windows.Forms.DialogResult.Cancel
-        Me.bnCancel.Location = New System.Drawing.Point(779, 3)
-        Me.bnCancel.Margin = New System.Windows.Forms.Padding(10, 0, 0, 0)
+        Me.bnCancel.Location = New System.Drawing.Point(1021, 3)
+        Me.bnCancel.Margin = New System.Windows.Forms.Padding(15, 0, 0, 0)
         Me.bnCancel.Size = New System.Drawing.Size(250, 70)
         Me.bnCancel.Text = "Cancel"
         '
@@ -161,18 +162,19 @@ Public Class eac3toForm
         '
         Me.bnOK.Anchor = System.Windows.Forms.AnchorStyles.None
         Me.bnOK.DialogResult = System.Windows.Forms.DialogResult.OK
-        Me.bnOK.Location = New System.Drawing.Point(518, 3)
-        Me.bnOK.Margin = New System.Windows.Forms.Padding(10, 0, 0, 0)
+        Me.bnOK.Location = New System.Drawing.Point(756, 3)
+        Me.bnOK.Margin = New System.Windows.Forms.Padding(15, 0, 0, 0)
         Me.bnOK.Size = New System.Drawing.Size(250, 70)
         Me.bnOK.Text = "OK"
         '
         'lvAudio
         '
         Me.lvAudio.Dock = System.Windows.Forms.DockStyle.Fill
+        Me.lvAudio.HideSelection = False
         Me.lvAudio.Location = New System.Drawing.Point(10, 0)
         Me.lvAudio.Margin = New System.Windows.Forms.Padding(10, 0, 10, 10)
         Me.lvAudio.Name = "lvAudio"
-        Me.lvAudio.Size = New System.Drawing.Size(997, 33)
+        Me.lvAudio.Size = New System.Drawing.Size(1251, 219)
         Me.lvAudio.TabIndex = 8
         Me.lvAudio.UseCompatibleStateImageBehavior = False
         '
@@ -181,10 +183,11 @@ Public Class eac3toForm
         Me.lvSubtitles.Anchor = CType((((System.Windows.Forms.AnchorStyles.Top Or System.Windows.Forms.AnchorStyles.Bottom) _
             Or System.Windows.Forms.AnchorStyles.Left) _
             Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
+        Me.lvSubtitles.HideSelection = False
         Me.lvSubtitles.Location = New System.Drawing.Point(10, 0)
         Me.lvSubtitles.Margin = New System.Windows.Forms.Padding(10, 0, 10, 10)
         Me.lvSubtitles.Name = "lvSubtitles"
-        Me.lvSubtitles.Size = New System.Drawing.Size(997, 24)
+        Me.lvSubtitles.Size = New System.Drawing.Size(1251, 148)
         Me.lvSubtitles.TabIndex = 9
         Me.lvSubtitles.UseCompatibleStateImageBehavior = False
         '
@@ -197,36 +200,36 @@ Public Class eac3toForm
         Me.flpSubtitleLinks.Controls.Add(Me.bnSubtitleNone)
         Me.flpSubtitleLinks.Controls.Add(Me.bnSubtitleEnglish)
         Me.flpSubtitleLinks.Controls.Add(Me.bnSubtitleNative)
-        Me.flpSubtitleLinks.Location = New System.Drawing.Point(10, 34)
+        Me.flpSubtitleLinks.Location = New System.Drawing.Point(10, 158)
         Me.flpSubtitleLinks.Margin = New System.Windows.Forms.Padding(10, 0, 0, 10)
         Me.flpSubtitleLinks.Name = "flpSubtitleLinks"
-        Me.flpSubtitleLinks.Size = New System.Drawing.Size(910, 70)
+        Me.flpSubtitleLinks.Size = New System.Drawing.Size(925, 70)
         Me.flpSubtitleLinks.TabIndex = 19
         '
         'bnSubtitleAll
         '
         Me.bnSubtitleAll.Location = New System.Drawing.Point(0, 0)
-        Me.bnSubtitleAll.Margin = New System.Windows.Forms.Padding(0, 0, 10, 0)
+        Me.bnSubtitleAll.Margin = New System.Windows.Forms.Padding(0, 0, 15, 0)
         Me.bnSubtitleAll.Size = New System.Drawing.Size(220, 70)
         Me.bnSubtitleAll.Text = "All"
         '
         'bnSubtitleNone
         '
-        Me.bnSubtitleNone.Location = New System.Drawing.Point(230, 0)
-        Me.bnSubtitleNone.Margin = New System.Windows.Forms.Padding(0, 0, 10, 0)
+        Me.bnSubtitleNone.Location = New System.Drawing.Point(235, 0)
+        Me.bnSubtitleNone.Margin = New System.Windows.Forms.Padding(0, 0, 15, 0)
         Me.bnSubtitleNone.Size = New System.Drawing.Size(220, 70)
         Me.bnSubtitleNone.Text = "None"
         '
         'bnSubtitleEnglish
         '
-        Me.bnSubtitleEnglish.Location = New System.Drawing.Point(460, 0)
-        Me.bnSubtitleEnglish.Margin = New System.Windows.Forms.Padding(0, 0, 10, 0)
+        Me.bnSubtitleEnglish.Location = New System.Drawing.Point(470, 0)
+        Me.bnSubtitleEnglish.Margin = New System.Windows.Forms.Padding(0, 0, 15, 0)
         Me.bnSubtitleEnglish.Size = New System.Drawing.Size(220, 70)
         Me.bnSubtitleEnglish.Text = "English"
         '
         'bnSubtitleNative
         '
-        Me.bnSubtitleNative.Location = New System.Drawing.Point(690, 0)
+        Me.bnSubtitleNative.Location = New System.Drawing.Point(705, 0)
         Me.bnSubtitleNative.Margin = New System.Windows.Forms.Padding(0)
         Me.bnSubtitleNative.Size = New System.Drawing.Size(220, 70)
         Me.bnSubtitleNative.Text = "Native"
@@ -240,36 +243,36 @@ Public Class eac3toForm
         Me.flpAudioLinks.Controls.Add(Me.bnAudioNone)
         Me.flpAudioLinks.Controls.Add(Me.bnAudioEnglish)
         Me.flpAudioLinks.Controls.Add(Me.bnAudioNative)
-        Me.flpAudioLinks.Location = New System.Drawing.Point(10, 43)
+        Me.flpAudioLinks.Location = New System.Drawing.Point(10, 229)
         Me.flpAudioLinks.Margin = New System.Windows.Forms.Padding(10, 0, 0, 0)
         Me.flpAudioLinks.Name = "flpAudioLinks"
-        Me.flpAudioLinks.Size = New System.Drawing.Size(910, 70)
+        Me.flpAudioLinks.Size = New System.Drawing.Size(925, 70)
         Me.flpAudioLinks.TabIndex = 18
         '
         'bnAudioAll
         '
         Me.bnAudioAll.Location = New System.Drawing.Point(0, 0)
-        Me.bnAudioAll.Margin = New System.Windows.Forms.Padding(0, 0, 10, 0)
+        Me.bnAudioAll.Margin = New System.Windows.Forms.Padding(0, 0, 15, 0)
         Me.bnAudioAll.Size = New System.Drawing.Size(220, 70)
         Me.bnAudioAll.Text = "All"
         '
         'bnAudioNone
         '
-        Me.bnAudioNone.Location = New System.Drawing.Point(230, 0)
-        Me.bnAudioNone.Margin = New System.Windows.Forms.Padding(0, 0, 10, 0)
+        Me.bnAudioNone.Location = New System.Drawing.Point(235, 0)
+        Me.bnAudioNone.Margin = New System.Windows.Forms.Padding(0, 0, 15, 0)
         Me.bnAudioNone.Size = New System.Drawing.Size(220, 70)
         Me.bnAudioNone.Text = "None"
         '
         'bnAudioEnglish
         '
-        Me.bnAudioEnglish.Location = New System.Drawing.Point(460, 0)
-        Me.bnAudioEnglish.Margin = New System.Windows.Forms.Padding(0, 0, 10, 0)
+        Me.bnAudioEnglish.Location = New System.Drawing.Point(470, 0)
+        Me.bnAudioEnglish.Margin = New System.Windows.Forms.Padding(0, 0, 15, 0)
         Me.bnAudioEnglish.Size = New System.Drawing.Size(220, 70)
         Me.bnAudioEnglish.Text = "English"
         '
         'bnAudioNative
         '
-        Me.bnAudioNative.Location = New System.Drawing.Point(690, 0)
+        Me.bnAudioNative.Location = New System.Drawing.Point(705, 0)
         Me.bnAudioNative.Margin = New System.Windows.Forms.Padding(0)
         Me.bnAudioNative.Size = New System.Drawing.Size(220, 70)
         Me.bnAudioNative.Text = "Native"
@@ -279,21 +282,21 @@ Public Class eac3toForm
         Me.cbVideoStream.Anchor = CType((System.Windows.Forms.AnchorStyles.Left Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
         Me.cbVideoStream.DropDownStyle = System.Windows.Forms.ComboBoxStyle.DropDownList
         Me.cbVideoStream.FormattingEnabled = True
-        Me.cbVideoStream.Location = New System.Drawing.Point(486, 3)
+        Me.cbVideoStream.Location = New System.Drawing.Point(546, 3)
         Me.cbVideoStream.Name = "cbVideoStream"
-        Me.cbVideoStream.Size = New System.Drawing.Size(540, 56)
+        Me.cbVideoStream.Size = New System.Drawing.Size(732, 56)
         Me.cbVideoStream.TabIndex = 16
         '
-        'Label8
+        'laStream
         '
-        Me.Label8.Anchor = System.Windows.Forms.AnchorStyles.None
-        Me.Label8.AutoSize = True
-        Me.Label8.Location = New System.Drawing.Point(341, 7)
-        Me.Label8.Name = "Label8"
-        Me.Label8.Size = New System.Drawing.Size(139, 48)
-        Me.Label8.TabIndex = 15
-        Me.Label8.Text = "Stream:"
-        Me.Label8.TextAlign = System.Drawing.ContentAlignment.MiddleLeft
+        Me.laStream.Anchor = System.Windows.Forms.AnchorStyles.None
+        Me.laStream.AutoSize = True
+        Me.laStream.Location = New System.Drawing.Point(381, 7)
+        Me.laStream.Name = "laStream"
+        Me.laStream.Size = New System.Drawing.Size(159, 48)
+        Me.laStream.TabIndex = 15
+        Me.laStream.Text = " Stream: "
+        Me.laStream.TextAlign = System.Drawing.ContentAlignment.MiddleLeft
         '
         'tlpTarget
         '
@@ -304,35 +307,36 @@ Public Class eac3toForm
         Me.tlpTarget.ColumnStyles.Add(New System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Percent, 50.0!))
         Me.tlpTarget.ColumnStyles.Add(New System.Windows.Forms.ColumnStyle())
         Me.tlpTarget.Controls.Add(Me.bnBrowse, 2, 0)
-        Me.tlpTarget.Controls.Add(Me.Label2, 0, 0)
+        Me.tlpTarget.Controls.Add(Me.laTargetDir, 0, 0)
         Me.tlpTarget.Controls.Add(Me.teTempDir, 1, 0)
-        Me.tlpTarget.Location = New System.Drawing.Point(10, 537)
-        Me.tlpTarget.Margin = New System.Windows.Forms.Padding(10, 0, 10, 0)
+        Me.tlpTarget.Location = New System.Drawing.Point(15, 847)
+        Me.tlpTarget.Margin = New System.Windows.Forms.Padding(15, 0, 15, 0)
         Me.tlpTarget.Name = "tlpTarget"
         Me.tlpTarget.RowCount = 1
         Me.tlpTarget.RowStyles.Add(New System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 50.0!))
-        Me.tlpTarget.Size = New System.Drawing.Size(1029, 79)
+        Me.tlpTarget.Size = New System.Drawing.Size(1271, 79)
         Me.tlpTarget.TabIndex = 4
         '
-        'Label2
+        'laTargetDir
         '
-        Me.Label2.Anchor = System.Windows.Forms.AnchorStyles.None
-        Me.Label2.AutoSize = True
-        Me.Label2.Location = New System.Drawing.Point(0, 15)
-        Me.Label2.Margin = New System.Windows.Forms.Padding(0)
-        Me.Label2.Name = "Label2"
-        Me.Label2.Size = New System.Drawing.Size(282, 48)
-        Me.Label2.TabIndex = 2
-        Me.Label2.Text = "Target Directory:"
-        Me.Label2.TextAlign = System.Drawing.ContentAlignment.MiddleLeft
+        Me.laTargetDir.Anchor = System.Windows.Forms.AnchorStyles.None
+        Me.laTargetDir.AutoSize = True
+        Me.laTargetDir.Location = New System.Drawing.Point(0, 15)
+        Me.laTargetDir.Margin = New System.Windows.Forms.Padding(0)
+        Me.laTargetDir.Name = "laTargetDir"
+        Me.laTargetDir.Size = New System.Drawing.Size(302, 48)
+        Me.laTargetDir.TabIndex = 2
+        Me.laTargetDir.Text = " Target Directory: "
+        Me.laTargetDir.TextAlign = System.Drawing.ContentAlignment.MiddleLeft
         '
         'teTempDir
         '
         Me.teTempDir.Anchor = CType((System.Windows.Forms.AnchorStyles.Left Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
-        Me.teTempDir.Location = New System.Drawing.Point(282, 4)
+        Me.teTempDir.BackColor = System.Drawing.Color.White
+        Me.teTempDir.Location = New System.Drawing.Point(302, 4)
         Me.teTempDir.Margin = New System.Windows.Forms.Padding(0, 0, 10, 0)
         Me.teTempDir.Name = "teTempDir"
-        Me.teTempDir.Size = New System.Drawing.Size(666, 70)
+        Me.teTempDir.Size = New System.Drawing.Size(859, 70)
         Me.teTempDir.TabIndex = 3
         '
         'tlpBottom
@@ -349,22 +353,22 @@ Public Class eac3toForm
         Me.tlpBottom.Controls.Add(Me.bnOK, 2, 0)
         Me.tlpBottom.Controls.Add(Me.bnMenu, 1, 0)
         Me.tlpBottom.Controls.Add(Me.cbChapters, 0, 0)
-        Me.tlpBottom.Location = New System.Drawing.Point(10, 626)
-        Me.tlpBottom.Margin = New System.Windows.Forms.Padding(10, 10, 10, 15)
+        Me.tlpBottom.Location = New System.Drawing.Point(15, 941)
+        Me.tlpBottom.Margin = New System.Windows.Forms.Padding(15)
         Me.tlpBottom.Name = "tlpBottom"
         Me.tlpBottom.RowCount = 1
         Me.tlpBottom.RowStyles.Add(New System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 100.0!))
-        Me.tlpBottom.Size = New System.Drawing.Size(1029, 76)
+        Me.tlpBottom.Size = New System.Drawing.Size(1271, 77)
         Me.tlpBottom.TabIndex = 20
         '
         'bnMenu
         '
         Me.bnMenu.Anchor = System.Windows.Forms.AnchorStyles.None
         Me.bnMenu.ContextMenuStrip = Me.cms
-        Me.bnMenu.Location = New System.Drawing.Point(437, 3)
+        Me.bnMenu.Location = New System.Drawing.Point(641, 3)
         Me.bnMenu.Margin = New System.Windows.Forms.Padding(0)
         Me.bnMenu.ShowMenuSymbol = True
-        Me.bnMenu.Size = New System.Drawing.Size(70, 70)
+        Me.bnMenu.Size = New System.Drawing.Size(100, 70)
         '
         'cms
         '
@@ -376,8 +380,8 @@ Public Class eac3toForm
         '
         Me.cbChapters.Anchor = System.Windows.Forms.AnchorStyles.Left
         Me.cbChapters.AutoSize = True
-        Me.cbChapters.Location = New System.Drawing.Point(5, 12)
-        Me.cbChapters.Margin = New System.Windows.Forms.Padding(5)
+        Me.cbChapters.Location = New System.Drawing.Point(15, 12)
+        Me.cbChapters.Margin = New System.Windows.Forms.Padding(15, 0, 0, 0)
         Me.cbChapters.Name = "cbChapters"
         Me.cbChapters.Size = New System.Drawing.Size(323, 52)
         Me.cbChapters.TabIndex = 2
@@ -395,7 +399,7 @@ Public Class eac3toForm
         Me.gbAudio.Margin = New System.Windows.Forms.Padding(10, 0, 10, 10)
         Me.gbAudio.Name = "gbAudio"
         Me.gbAudio.Padding = New System.Windows.Forms.Padding(5)
-        Me.gbAudio.Size = New System.Drawing.Size(1027, 263)
+        Me.gbAudio.Size = New System.Drawing.Size(1281, 449)
         Me.gbAudio.TabIndex = 23
         Me.gbAudio.TabStop = False
         Me.gbAudio.Text = "Audio"
@@ -415,7 +419,7 @@ Public Class eac3toForm
         Me.tlpAudio.RowStyles.Add(New System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 100.0!))
         Me.tlpAudio.RowStyles.Add(New System.Windows.Forms.RowStyle())
         Me.tlpAudio.RowStyles.Add(New System.Windows.Forms.RowStyle())
-        Me.tlpAudio.Size = New System.Drawing.Size(1017, 205)
+        Me.tlpAudio.Size = New System.Drawing.Size(1271, 391)
         Me.tlpAudio.TabIndex = 23
         '
         'tlpAudioOptions
@@ -427,45 +431,45 @@ Public Class eac3toForm
         Me.tlpAudioOptions.ColumnStyles.Add(New System.Windows.Forms.ColumnStyle())
         Me.tlpAudioOptions.ColumnStyles.Add(New System.Windows.Forms.ColumnStyle())
         Me.tlpAudioOptions.ColumnStyles.Add(New System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Percent, 100.0!))
-        Me.tlpAudioOptions.Controls.Add(Me.Label1, 0, 0)
-        Me.tlpAudioOptions.Controls.Add(Me.Label3, 2, 0)
+        Me.tlpAudioOptions.Controls.Add(Me.laOutput, 0, 0)
+        Me.tlpAudioOptions.Controls.Add(Me.laOptions, 2, 0)
         Me.tlpAudioOptions.Controls.Add(Me.cbAudioOutput, 1, 0)
         Me.tlpAudioOptions.Controls.Add(Me.cmdlOptions, 3, 0)
-        Me.tlpAudioOptions.Location = New System.Drawing.Point(3, 116)
+        Me.tlpAudioOptions.Location = New System.Drawing.Point(3, 302)
         Me.tlpAudioOptions.Name = "tlpAudioOptions"
         Me.tlpAudioOptions.RowCount = 1
         Me.tlpAudioOptions.RowStyles.Add(New System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 100.0!))
-        Me.tlpAudioOptions.Size = New System.Drawing.Size(1011, 86)
+        Me.tlpAudioOptions.Size = New System.Drawing.Size(1265, 86)
         Me.tlpAudioOptions.TabIndex = 22
         '
-        'Label1
+        'laOutput
         '
-        Me.Label1.Anchor = System.Windows.Forms.AnchorStyles.None
-        Me.Label1.AutoSize = True
-        Me.Label1.Location = New System.Drawing.Point(3, 19)
-        Me.Label1.Name = "Label1"
-        Me.Label1.Size = New System.Drawing.Size(140, 48)
-        Me.Label1.TabIndex = 19
-        Me.Label1.Text = "Output:"
-        Me.Label1.TextAlign = System.Drawing.ContentAlignment.MiddleLeft
+        Me.laOutput.Anchor = System.Windows.Forms.AnchorStyles.None
+        Me.laOutput.AutoSize = True
+        Me.laOutput.Location = New System.Drawing.Point(3, 19)
+        Me.laOutput.Name = "laOutput"
+        Me.laOutput.Size = New System.Drawing.Size(160, 48)
+        Me.laOutput.TabIndex = 19
+        Me.laOutput.Text = " Output: "
+        Me.laOutput.TextAlign = System.Drawing.ContentAlignment.MiddleLeft
         '
-        'Label3
+        'laOptions
         '
-        Me.Label3.Anchor = System.Windows.Forms.AnchorStyles.None
-        Me.Label3.AutoSize = True
-        Me.Label3.Location = New System.Drawing.Point(361, 19)
-        Me.Label3.Name = "Label3"
-        Me.Label3.Size = New System.Drawing.Size(153, 48)
-        Me.Label3.TabIndex = 20
-        Me.Label3.Text = "Options:"
-        Me.Label3.TextAlign = System.Drawing.ContentAlignment.MiddleLeft
+        Me.laOptions.Anchor = System.Windows.Forms.AnchorStyles.None
+        Me.laOptions.AutoSize = True
+        Me.laOptions.Location = New System.Drawing.Point(381, 19)
+        Me.laOptions.Name = "laOptions"
+        Me.laOptions.Size = New System.Drawing.Size(173, 48)
+        Me.laOptions.TabIndex = 20
+        Me.laOptions.Text = " Options: "
+        Me.laOptions.TextAlign = System.Drawing.ContentAlignment.MiddleLeft
         '
         'cbAudioOutput
         '
         Me.cbAudioOutput.Anchor = System.Windows.Forms.AnchorStyles.None
         Me.cbAudioOutput.DropDownStyle = System.Windows.Forms.ComboBoxStyle.DropDownList
         Me.cbAudioOutput.FormattingEnabled = True
-        Me.cbAudioOutput.Location = New System.Drawing.Point(149, 20)
+        Me.cbAudioOutput.Location = New System.Drawing.Point(169, 15)
         Me.cbAudioOutput.Name = "cbAudioOutput"
         Me.cbAudioOutput.Size = New System.Drawing.Size(206, 56)
         Me.cbAudioOutput.TabIndex = 21
@@ -476,11 +480,11 @@ Public Class eac3toForm
             Or System.Windows.Forms.AnchorStyles.Left) _
             Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
         Me.gbSubtitles.Controls.Add(Me.tlpSubtitles)
-        Me.gbSubtitles.Location = New System.Drawing.Point(10, 355)
+        Me.gbSubtitles.Location = New System.Drawing.Point(10, 541)
         Me.gbSubtitles.Margin = New System.Windows.Forms.Padding(10, 0, 10, 10)
         Me.gbSubtitles.Name = "gbSubtitles"
         Me.gbSubtitles.Padding = New System.Windows.Forms.Padding(5)
-        Me.gbSubtitles.Size = New System.Drawing.Size(1027, 172)
+        Me.gbSubtitles.Size = New System.Drawing.Size(1281, 296)
         Me.gbSubtitles.TabIndex = 24
         Me.gbSubtitles.TabStop = False
         Me.gbSubtitles.Text = "Subtitles"
@@ -498,7 +502,7 @@ Public Class eac3toForm
         Me.tlpSubtitles.RowCount = 2
         Me.tlpSubtitles.RowStyles.Add(New System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 50.0!))
         Me.tlpSubtitles.RowStyles.Add(New System.Windows.Forms.RowStyle())
-        Me.tlpSubtitles.Size = New System.Drawing.Size(1017, 114)
+        Me.tlpSubtitles.Size = New System.Drawing.Size(1271, 238)
         Me.tlpSubtitles.TabIndex = 20
         '
         'tlpVideo
@@ -515,14 +519,14 @@ Public Class eac3toForm
         Me.tlpVideo.ColumnStyles.Add(New System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Percent, 100.0!))
         Me.tlpVideo.Controls.Add(Me.cbVideoStream, 3, 0)
         Me.tlpVideo.Controls.Add(Me.cbVideoOutput, 1, 0)
-        Me.tlpVideo.Controls.Add(Me.Label8, 2, 0)
-        Me.tlpVideo.Controls.Add(Me.Label5, 0, 0)
+        Me.tlpVideo.Controls.Add(Me.laStream, 2, 0)
+        Me.tlpVideo.Controls.Add(Me.laVideo, 0, 0)
         Me.tlpVideo.Location = New System.Drawing.Point(10, 10)
         Me.tlpVideo.Margin = New System.Windows.Forms.Padding(10)
         Me.tlpVideo.Name = "tlpVideo"
         Me.tlpVideo.RowCount = 1
         Me.tlpVideo.RowStyles.Add(New System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 100.0!))
-        Me.tlpVideo.Size = New System.Drawing.Size(1029, 62)
+        Me.tlpVideo.Size = New System.Drawing.Size(1281, 62)
         Me.tlpVideo.TabIndex = 25
         '
         'tlpMain
@@ -543,7 +547,7 @@ Public Class eac3toForm
         Me.tlpMain.RowStyles.Add(New System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 40.0!))
         Me.tlpMain.RowStyles.Add(New System.Windows.Forms.RowStyle())
         Me.tlpMain.RowStyles.Add(New System.Windows.Forms.RowStyle())
-        Me.tlpMain.Size = New System.Drawing.Size(1049, 717)
+        Me.tlpMain.Size = New System.Drawing.Size(1301, 1033)
         Me.tlpMain.TabIndex = 26
         '
         'eac3toForm
@@ -552,14 +556,14 @@ Public Class eac3toForm
         Me.AutoScaleDimensions = New System.Drawing.SizeF(288.0!, 288.0!)
         Me.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Dpi
         Me.CancelButton = Me.bnCancel
-        Me.ClientSize = New System.Drawing.Size(1049, 717)
+        Me.ClientSize = New System.Drawing.Size(1301, 1033)
         Me.Controls.Add(Me.tlpMain)
         Me.FormBorderStyle = System.Windows.Forms.FormBorderStyle.FixedSingle
         Me.HelpButton = False
         Me.KeyPreview = True
-        Me.Margin = New System.Windows.Forms.Padding(10, 10, 10, 10)
+        Me.Margin = New System.Windows.Forms.Padding(10)
         Me.Name = "eac3toForm"
-        Me.Text = "eac3to"
+        Me.Text = "eac3to Demuxing"
         Me.flpSubtitleLinks.ResumeLayout(False)
         Me.flpAudioLinks.ResumeLayout(False)
         Me.tlpTarget.ResumeLayout(False)
@@ -591,7 +595,9 @@ Public Class eac3toForm
 
     Private Output As String
     Private Streams As New BindingList(Of M2TSStream)
-    Private AudioOutputFormats As String() = {"m4a", "ac3", "dts", "flac", "wav", "dtsma", "dtshr", "eac3", "thd", "thd+ac3"}
+    Private AudioOutputFormats As String() =
+        {"m4a", "ac3", "dts", "flac", "wav", "dtsma", "dtshr", "eac3", "thd", "thd+ac3"}
+
     Private Project As Project
 
     Sub New(proj As Project)
@@ -655,8 +661,8 @@ Public Class eac3toForm
         form.Doc.WriteStart("Audio Stream Profiles")
         form.Doc.WriteParagraph("Allows to automatically apply default values for audio streams.")
         form.Doc.WriteTable({New StringPair("Match All", "space separated, if all match then the Output Format and Options are applied"),
-                          New StringPair("Output Format", "applied to the stream if Match All succeeds"),
-                          New StringPair("Options", "applied to the stream if Match All succeeds")})
+                             New StringPair("Output Format", "applied to the stream if Match All succeeds"),
+                             New StringPair("Options", "applied to the stream if Match All succeeds")})
         form.Show()
     End Sub
 
@@ -818,7 +824,7 @@ Public Class eac3toForm
                         Next
 
                         Select Case ms.Codec
-                            Case "AC3 EX", "AC3 Surround", "AC3 Headphone"
+                            Case "AC3", "AC3 EX", "AC3 Surround", "AC3 Headphone"
                                 ms.OutputType = "ac3"
                             Case "E-AC3", "E-AC3 EX"
                                 ms.OutputType = "eac3"
@@ -834,20 +840,24 @@ Public Class eac3toForm
                                 ms.OutputType = "dtshr"
                             Case "RAW/PCM"
                                 ms.OutputType = "flac"
+                            Case "AAC"
+                                ms.OutputType = "m4a"
+                            Case "Subtitle (ASS)"
+                                ms.OutputType = "ass"
                             Case Else
                                 ms.OutputType = ms.Codec.ToLower.Replace("-", "")
                         End Select
                     End If
 
-                    For Each iProfile In s.eac3toProfiles
-                        Dim searchWords = iProfile.Match.SplitNoEmptyAndWhiteSpace(" ")
+                    For Each pro In s.eac3toProfiles
+                        Dim searchWords = pro.Match.SplitNoEmptyAndWhiteSpace(" ")
                         If searchWords.NothingOrEmpty Then
                             Continue For
                         End If
 
                         If ms.Text.ContainsAll(searchWords) Then
-                            ms.OutputType = iProfile.Output
-                            ms.Options = iProfile.Options
+                            ms.OutputType = pro.Output
+                            ms.Options = pro.Options
                         End If
                     Next
 
@@ -1076,65 +1086,65 @@ Public Class eac3toForm
     End Sub
 
     Sub bnAudioAll_Click(sender As Object, e As EventArgs) Handles bnAudioAll.Click
-        For Each i As ListViewItem In lvAudio.Items
-            i.Checked = True
+        For Each item As ListViewItem In lvAudio.Items
+            item.Checked = True
         Next
     End Sub
 
     Sub bnAudioNone_Click(sender As Object, e As EventArgs) Handles bnAudioNone.Click
-        For Each i As ListViewItem In lvAudio.Items
-            i.Checked = False
+        For Each item As ListViewItem In lvAudio.Items
+            item.Checked = False
         Next
     End Sub
 
     Sub bnAudioEnglish_Click(sender As Object, e As EventArgs) Handles bnAudioEnglish.Click
-        For Each i As ListViewItem In lvAudio.Items
-            Dim stream = DirectCast(i.Tag, M2TSStream)
+        For Each item As ListViewItem In lvAudio.Items
+            Dim stream = DirectCast(item.Tag, M2TSStream)
 
             If stream.Language.TwoLetterCode = "en" Then
-                i.Checked = True
+                item.Checked = True
             End If
         Next
     End Sub
 
     Sub bnAudioNative_Click(sender As Object, e As EventArgs) Handles bnAudioNative.Click
-        For Each i As ListViewItem In lvAudio.Items
-            Dim stream = DirectCast(i.Tag, M2TSStream)
+        For Each item As ListViewItem In lvAudio.Items
+            Dim stream = DirectCast(item.Tag, M2TSStream)
 
             If stream.Language.TwoLetterCode = CultureInfo.CurrentCulture.TwoLetterISOLanguageName Then
-                i.Checked = True
+                item.Checked = True
             End If
         Next
     End Sub
 
     Sub bnSubtitleAll_Click(sender As Object, e As EventArgs) Handles bnSubtitleAll.Click
-        For Each i As ListViewItem In lvSubtitles.Items
-            i.Checked = True
+        For Each item As ListViewItem In lvSubtitles.Items
+            item.Checked = True
         Next
     End Sub
 
     Sub bnSubtitleNone_Click(sender As Object, e As EventArgs) Handles bnSubtitleNone.Click
-        For Each i As ListViewItem In lvSubtitles.Items
-            i.Checked = False
+        For Each item As ListViewItem In lvSubtitles.Items
+            item.Checked = False
         Next
     End Sub
 
     Sub bnSubtitleEnglish_Click(sender As Object, e As EventArgs) Handles bnSubtitleEnglish.Click
-        For Each i As ListViewItem In lvSubtitles.Items
-            Dim stream = DirectCast(i.Tag, M2TSStream)
+        For Each item As ListViewItem In lvSubtitles.Items
+            Dim stream = DirectCast(item.Tag, M2TSStream)
 
             If stream.Language.TwoLetterCode = "en" Then
-                i.Checked = True
+                item.Checked = True
             End If
         Next
     End Sub
 
     Sub bnSubtitleNative_Click(sender As Object, e As EventArgs) Handles bnSubtitleNative.Click
-        For Each i As ListViewItem In lvSubtitles.Items
-            Dim stream = DirectCast(i.Tag, M2TSStream)
+        For Each item As ListViewItem In lvSubtitles.Items
+            Dim stream = DirectCast(item.Tag, M2TSStream)
 
             If stream.Language.TwoLetterCode = CultureInfo.CurrentCulture.TwoLetterISOLanguageName Then
-                i.Checked = True
+                item.Checked = True
             End If
         Next
     End Sub
@@ -1148,11 +1158,11 @@ Public Class eac3toForm
 
         Dim hdCounter As Integer
 
-        For Each i In Streams
-            If i.Checked AndAlso i.IsAudio Then
-                If {"dtsma", "dtshr"}.Contains(i.OutputType) Then
+        For Each str In Streams
+            If str.Checked AndAlso str.IsAudio Then
+                If {"dtsma", "dtshr"}.Contains(str.OutputType) Then
                     hdCounter += 1
-                ElseIf i.OutputType = "dts" AndAlso {"DTS Master Audio", "DTS Hi-Res"}.Contains(i.Codec) Then
+                ElseIf str.OutputType = "dts" AndAlso {"DTS Master Audio", "DTS Hi-Res"}.Contains(str.Codec) Then
                     hdCounter -= 1
                 End If
             End If
@@ -1176,7 +1186,7 @@ Public Class eac3toForm
     End Sub
 
     Protected Overrides Sub OnShown(e As EventArgs)
-        Task.Run(AddressOf StartAnalyze)
         MyBase.OnShown(e)
+        Task.Run(AddressOf StartAnalyze)
     End Sub
 End Class